==> R/01_packages.R <==
# Packages
library(tidyverse)
library(janitor)
library(lubridate)
library(readxl)
library(rvest)
library(stringi)
library(data.table)
library(quanteda)
library(quanteda.textstats)
library(quanteda.textmodels)
library(sf)
library(stm)
library(tidystm)
library(ggraph)
library(webshot2)
library(tidytext)
library(flextable)
library(ggrepel)
library(scales)
library(furrr)
library(progressr)
library(future)
library(xlsx)
